requirements.txt
dotmap==1.3.24
ftfy==6.0.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
PyYAML==6.0
RandAugment==1.0.2
regex==2021.10.23
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
wandb==0.12.5
